+++ cmd/complex_client/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"log"
 	"math/big"
 	"net/http"
@@ -19,35 +18,30 @@ func main() {
 	defer cancel()
 
 	r, _ := http.NewRequestWithContext(ctx, http.MethodGet, "http://localhost:8080/events", nil)
-	conn, done, events := sse.NewConnection(r), make(chan struct{}), make(chan sse.Event, 1)
+	conn := sse.NewConnection(r)
+	// Callbacks are called from separate goroutines, we must synchronize access to stdout.
+	// log.Logger does that automatically for us, so we create one that writes to stdout without any prefix or flags.
+	out := log.New(os.Stdout, "", 0)
 
-	conn.SubscribeToAll(events)
-
-	go func() {
-		for event := range events {
-			switch event.Name {
-			case "cycles", "ops":
-				fmt.Printf("Metric %s: %s\n", event.Name, event)
-			case "close":
-				fmt.Println("Server closed!")
-				cancel()
-			default: // no event name
-				var sum, num big.Int
-				for _, n := range strings.Split(event.String(), "\n") {
-					_, _ = num.SetString(n, 10)
-					sum.Add(&sum, &num)
-				}
-
-				fmt.Printf("Sum of random numbers: %s\n", &sum)
+	conn.SubscribeToAll(func(event sse.Event) {
+		switch event.Name {
+		case "cycles", "ops":
+			out.Printf("Metric %s: %s\n", event.Name, event)
+		case "close":
+			out.Println("Server closed!")
+			cancel()
+		default: // no event name
+			var sum, num big.Int
+			for _, n := range strings.Split(event.String(), "\n") {
+				_, _ = num.SetString(n, 10)
+				sum.Add(&sum, &num)
 			}
-		}
 
-		close(done)
-	}()
+			out.Printf("Sum of random numbers: %s\n", &sum)
+		}
+	})
 
 	if err := conn.Connect(); err != nil {
-		log.Println(err)
+		out.Println(err)
 	}
-
-	<-done
 }