+++ cmd/helloworld_client/client.go
@@ -1,9 +1,9 @@
 package main
 
 import (
-	"fmt"
 	"log"
 	"net/http"
+	"os"
 
 	"github.com/tmaxmax/go-sse"
 )
@@ -11,17 +11,15 @@ import (
 func main() {
 	r, _ := http.NewRequest(http.MethodGet, "http://localhost:8000", nil)
 	conn := sse.NewConnection(r)
-	ch := make(chan sse.Event)
+	// Callbacks are called from separate goroutines, we must synchronize access to stdout.
+	// log.Logger does that automatically for us, so we create one that writes to stdout without any prefix or flags.
+	out := log.New(os.Stdout, "", 0)
 
-	conn.SubscribeMessages(ch)
-
-	go func() {
-		for ev := range ch {
-			fmt.Printf("%s\n\n", ev.Data)
-		}
-	}()
+	conn.SubscribeMessages(func(event sse.Event) {
+		out.Printf("%s\n\n", event)
+	})
 
 	if err := conn.Connect(); err != nil {
-		log.Println(err)
+		out.Println(err)
 	}
 }